+++ class/actions_subtotal.class.php
@@ -2966,9 +2966,24 @@ class ActionsSubtotal
 						 }
 						if (TSubtotal::isTitle($line)) {
 							//Folder for expand
-							print ' <a class="collapse_bom" id="collapse-' . $line->id . '" href="#">';
+							$titleAttr = ($line->array_options['options_hideblock'] == 1) ? $langs->trans("Subtotal_Show") : $langs->trans("Subtotal_Hide");
+
+							print '<span class="fold-subtotal-container" >';
+
+							// bouton pour ouvrir/fermer le bloc
+							print ' <span title="'.dol_escape_htmltag($titleAttr).'" class="fold-subtotal-btn" data-toggle-all-children="0" data-title-line-target="' . $line->id . '" id="collapse-' . $line->id . '" >';
 							print (($line->array_options['options_hideblock'] == 1) ? img_picto('', 'folder') : img_picto('', 'folder-open'));
-							print '</a>';
+							print '</span>';
+
+							// Bouton pour ouvrir/fermer aussi les enfants
+							print ' <span title="'.dol_escape_htmltag($titleAttr).'" class="fold-subtotal-btn" data-toggle-all-children="1" data-title-line-target="' . $line->id . '" id="collapse-children-' . $line->id . '" >';
+							print (($line->array_options['options_hideblock'] == 1) ? img_picto('', 'folder') : img_picto('', 'folder-open'));
+							print '</span>';
+
+							// un span pour contenir des infos comme le nombre de lignes cachées etc...
+							print ' <span class="fold-subtotal-info" title="'.dol_escape_htmltag($langs->trans('NumberOfHiddenLines')).'" data-title-line-target="' . $line->id . '" ></span>';
+
+							print '</span>';
 						}
 
 
@@ -3579,8 +3594,11 @@ class ActionsSubtotal
 				$TSubNc[$l->id] = (int) $l->array_options['options_subtotal_nc'];
 			}
 
+			print '<script type="text/javascript" src="'.dol_buildpath('subtotal/js/subtotal.lib.js', 1).'"></script>';
+
 			$form = new Form($db);
 			?>
+
 			<script type="text/javascript">
 				$(function() {
 					var subtotal_TSubNc = <?php echo json_encode($TSubNc); ?>;
@@ -3710,13 +3728,17 @@ class ActionsSubtotal
 
 	    if(TSubtotal::isTitle($line)){
 	        $ThtmlData['data-issubtotal'] = 'title';
-	    }elseif(TSubtotal::isSubtotal($line)){
+
+			$ThtmlData['data-folder-status'] = 'open';
+			if(!empty($line->array_options['options_hideblock'])){
+				$ThtmlData['data-folder-status'] = 'closed';
+			}
+		}elseif(TSubtotal::isSubtotal($line)){
 	        $ThtmlData['data-issubtotal'] = 'subtotal';
 	    }
 	    else{
 	        $ThtmlData['data-issubtotal'] = 'freetext';
 	    }
-
 
 	    // Change or add data  from hooks
 	    $parameters = array_replace($parameters , array(  'ThtmlData' => $ThtmlData )  );
@@ -3776,8 +3798,9 @@ class ActionsSubtotal
 
 			$jsConf = array(
 				'useOldSplittedTrForLine' => intval(DOL_VERSION) < 16 ? 1 : 0,
-			)
+			);
 
+			print '<script type="text/javascript" src="'.dol_buildpath('subtotal/js/subtotal.lib.js', 1).'"></script>';
 	        ?>
 
 
@@ -3877,57 +3900,6 @@ class ActionsSubtotal
 			    	    }
 			    });
  				<?php } ?>
-
-				function getSubtotalTitleChilds(item)
-				{
-		    		var TcurrentChilds = []; // = JSON.parse(item.attr('data-childrens'));
-		    		var level = item.data('level');
-
-		    		var indexOfFirstSubtotal = -1;
-		    		var indexOfFirstTitle = -1;
-
-		    		item.nextAll('[id^="row-"]').each(function(index){
-
-						var dataLevel = $(this).attr('data-level');
-						var dataIsSubtotal = $(this).attr('data-issubtotal');
-
-						if(dataIsSubtotal != 'undefined' && dataLevel != 'undefined' )
-						{
-
-							if(dataLevel <=  level && indexOfFirstSubtotal < 0 && dataIsSubtotal == 'subtotal' )
-							{
-								indexOfFirstSubtotal = index;
-								if(indexOfFirstTitle < 0)
-								{
-									TcurrentChilds.push($(this).attr('id'));
-								}
-							}
-
-							if(dataLevel <=  level && indexOfFirstSubtotal < 0 && indexOfFirstTitle < 0 && dataIsSubtotal == 'title' )
-							{
-								indexOfFirstTitle = index;
-							}
-						}
-
-						if(indexOfFirstTitle < 0 && indexOfFirstSubtotal < 0)
-						{
-							TcurrentChilds.push($(this).attr('id'));
-
-							// Add extraffield support for dolibarr > 7
-							var thisId = $(this).attr('data-id');
-							var thisElement = $(this).attr('data-element');
-							if(thisId != undefined && thisElement != undefined && subTotalConf.useOldSplittedTrForLine )
-							{
-								$('[data-targetid="' + thisId + '"][data-element="extrafield"][data-targetelement="'+ thisElement +'"]').each(function(index){
-									TcurrentChilds.push($(this).attr('id'));
-								});
-							}
-
-						}
-
-		    		});
-		    		return TcurrentChilds;
-				}
 
 			});
 			</script>
@@ -4050,7 +4022,7 @@ class ActionsSubtotal
 
 	public function printCommonFooter(&$parameters, &$object, &$action, $hookmanager)
 	{
-			global $langs, $db;
+			global $langs, $db, $conf;
 
 			$contextArray = explode(':',$parameters['context']);
 
@@ -4088,211 +4060,452 @@ class ActionsSubtotal
 						if ($line->array_options['options_hideblock']) $TBlocksToHide[] = $line->id;
 					}
 				}
-					?>
-						<script type="text/javascript">
-							$(document).ready(function(){
 
-								//A chaque chargement de la page, on cache toutes les lignes qui font parties des blocs à cacher (sauf les titres et sous-totaux)
-								<?php if(!empty($TBlocksToHide)) { foreach($TBlocksToHide as $id_linetitle){ ?>
-								var element = $("#collapse-<?php echo $id_linetitle ?>");
-								folderManage(element);
-								<?php }} ?>
 
-								//Lors du clic sur un dossier, on cache ou faire apparaitre les lignes contenues dans le bloc concerné
-								$(".collapse_bom").click(function() {
-									folderManage_click($(this));
-									return false;
-								});
+				$hideMode = !empty($conf->global->SUBTOTAL_BLOC_FOLD_MODE)?$conf->global->SUBTOTAL_BLOC_FOLD_MODE:'default';
+				if(!in_array($hideMode, array('default', 'keepTitle'))){
+					$hideMode = 'default';
+				}
 
-								allFolderManage();
 
-								//Fonction qui permet la gestion de l'affichage des blocs à chaque chargement de page
-								function folderManage(element){
-									//On récupère le titre concerné
-									var id_line_title = element.attr('id').replace('collapse-', '');
+				$jsConf = array(
+					'linesToHide' => $TBlocksToHide,
+					'closeMode' => $hideMode, // default, keepTitle
+					'interfaceUrl'=> dol_buildpath('/subtotal/script/interface.php', 1),
+					'element' => $element,
+					'element_id' => $id,
+					'img_folder_closed' => img_picto('', 'folder'),
+					'img_folder_open' => img_picto('', 'folder-open'),
+					'langs' => array(
+						'Subtotal_HideAll' => $langs->transnoentities("Subtotal_HideAll"),
+						'Subtotal_ShowAll' => $langs->transnoentities("Subtotal_ShowAll"),
+						'Subtotal_Hide' => $langs->transnoentities("Subtotal_Hide"),
+						'Subtotal_Show' => $langs->transnoentities("Subtotal_Show"),
+						'Subtotal_ForceHideAll' => $langs->transnoentities("Subtotal_ForceHideAll"),
+						'Subtotal_ForceShowAll' => $langs->transnoentities("Subtotal_ForceShowAll")
+					)
+				);
 
-									//on récupère les lignes concernées en fonction du titre
-									$.ajax({
-										url: '<?php echo dol_buildpath('/subtotal/script/interface.php', 1); ?>'
-										,type: 'POST'
-										,data: {
-											json:1
-											,get: 'getLinesFromTitle'
-											,element: '<?php echo $element; ?>'
-											,elementid: '<?php echo $id; ?>'
-											,lineid: id_line_title
+				print '<script type="text/javascript" src="'.dol_buildpath('subtotal/js/subtotal.lib.js', 1).'"></script>';
+
+				?>
+				<style>
+					.fold-subtotal-container{
+						-webkit-user-select: none; /* Safari */
+						-ms-user-select: none; /* IE 10 and IE 11 */
+						user-select: none; /* Standard syntax */
+					}
+
+					.toggle-all-folder-status, .fold-subtotal-btn{
+						cursor: pointer;
+					}
+					.fold-subtotal-btn[data-toggle-all-children="1"]{
+						color: rgb(190, 53, 53);
+					}
+					.toggle-all-folder-status:hover, .fold-subtotal-btn:hover{
+						color: var(--colortextlink, rgb(10, 20, 100));
+					}
+					.fold-subtotal-btn[data-toggle-all-children="1"]:hover{
+						color: rgb(138, 28, 28);
+					}
+				</style>
+				<script type="text/javascript">
+					// TODO : mettre ça dans une classe js
+					$(document).ready(function(){
+						// Utilisation d'une sorte de namespace en JS
+						subtotalFolders = {};
+						(function(o) {
+							o.config = <?php print json_encode($jsConf); ?> ;
+
+							/**
+							 * Dolibarr token
+							 * @type {string}
+							 */
+							o.newToken = '';
+
+							/**
+							 *
+							 * @param {int} titleId
+							 */
+							o.countHiddenLinesForTitle = function(titleId){
+								let $titleLine = $('#row-' + titleId);
+								let childrenList = getSubtotalTitleChilds($titleLine, true); // renvoi la liste des id des enfants
+								let totalHiddenLines = 0;
+								if(childrenList.length > 0) {
+									childrenList.forEach((childLineId) => {
+										let $childLine = $('#'+childLineId);
+										if(!$childLine.is(":visible")){
+											totalHiddenLines++;
 										}
-									}).done(function(data) {
-
-										let TSubLines = [];
-										var data = JSON.parse(data);
-										//en fonction des lignes concernées, on définit les #id (html) concernés
-										$.each(data, function (title, tlines) {
-											$.each(tlines, function (key, line) {
-												TSubLines.push('#row-' + line);
-											});
-										});
-
-										//pour chaque #id (html) concerné, on cache la ligne
-										$.each(TSubLines, function(key, value){
-											$(value).hide();
-										});
 									});
-
-									//on met à jour l'icône "dossier" pour indiquer qu'il est fermé
-									element.html('<?php echo dol_escape_js(img_picto('', 'folder')); ?>');
 								}
 
+								return totalHiddenLines;
+							}
 
-								//Fonction qui permet la gestion des blocs cachés ou non lors du clic sur l'icône "dossier"
-								function folderManage_click(element) {
-									//On récupère le titre concerné
-									var id_line_title = element.attr('id').replace('collapse-', '');
+							/**
+							 * Mise à jour des titres parents pour l'affichage du nombre de lignes cachées
+							 * @param {jQuery}  $childTilteLine la ligne de titre enfant
+							 */
+							o.updateHiddenLinesCountInfoForParentTitles = function($childTilteLine){
+								let parentTitles = o.getTitleParents($childTilteLine);
+								if(parentTitles.length>0){
+									parentTitles.forEach((parentTitleLineId) => {
+										let $titleCollapseInfos = $('.fold-subtotal-info[data-title-line-target="' + parentTitleLineId + '"]');
+										if($titleCollapseInfos.length>0){
+											let totalHiddenLines = o.countHiddenLinesForTitle(parentTitleLineId);
+											$titleCollapseInfos.html('('+ totalHiddenLines +')');
+											if(totalHiddenLines == 0){
+												$titleCollapseInfos.html('');
+											}
+										}
+									});
+								}
+							}
 
-									//en fonction de l'état de l'icône "folder" on détermine si l'action demandée est de caché le dossier ou de montrer le dossier
-									if(element.html().indexOf('folder-open') <= 0){
-										var action_hide = 0;
-									} else {
-										var action_hide = 1;
+							/**
+							 * @param {jQuery}  $childLine
+							 * @param {int} titleId
+							 */
+							o.addTitleParentId = function($childLine, titleId){
+								// Ajoute l'id parent si se n'est pas déja fait
+								let parentTitleIds = $childLine.attr('data-parent-titles');
+								if(parentTitleIds != null){
+									let parentTitleIdsList = parentTitleIds.split(",");
+									if(!parentTitleIdsList.includes(titleId)){
+										$childLine.attr('data-parent-titles', parentTitleIds + ',' + titleId);
+									}
+								}else{
+									$childLine.attr('data-parent-titles', titleId);
+								}
+							}
+
+							/**
+							 * @param {jQuery}  $childLine
+							 * @param {int} titleId
+							 * @return []
+							 */
+							o.getTitleParents = function($childLine){
+								let result = [];
+								let parentTitleIds = $childLine.attr('data-parent-titles');
+								if(parentTitleIds != null){
+									return parentTitleIds.split(",");
+								}
+
+								return result;
+							}
+
+							/**
+							 *
+							 * @param {int} titleId
+							 * @param toggleStatus : open, closed
+							 */
+							o.toggleChildFolderStatusDisplay = function(titleId, toggleStatus = 'open'){
+								let $titleLine = $('#row-' + titleId);
+								let $collapseBtn = $('.fold-subtotal-btn[data-title-line-target="' + titleId + '"]');
+								let $collapseSimpleBtn = $('.fold-subtotal-btn[data-title-line-target="' + titleId + '"][data-toggle-all-children="0"]');
+								let $collapseAllBtn = $('.fold-subtotal-btn[data-title-line-target="' + titleId + '"][data-toggle-all-children="1"]');
+								let $collapseInfos = $('.fold-subtotal-info[data-title-line-target="' + titleId + '"]');
+
+								if($titleLine.length>0){
+									$titleLine.attr('data-folder-status', toggleStatus);
+									let haveTitle = false;
+									let childrenList = getSubtotalTitleChilds($titleLine, true); // renvoi la liste des id des enfants
+									let totalHiddenLines = 0;
+
+									if(childrenList.length > 0) {
+
+										let doNotDisplayLines = []; // Dans le cas de l'ouverture il faut vérifier que les titres enfants ne sont pas fermés avant d'ouvrir
+										let doNotHiddeLines = []; // En mode keepTitle: Dans le cas de la fermeture il faut vérifier que les titres enfants ne sont pas ouvert avant de fermer
+
+										childrenList.forEach((childLineId) => {
+											let $childLine = $('#'+childLineId);
+
+											if ($childLine.attr('data-issubtotal') == "title"){
+
+												// Ajoute l'id parent si se n'est pas déja fait
+												o.addTitleParentId($childLine, titleId);
+
+												haveTitle = true;
+												// Dans le cas de l'ouverture il faut vérifier que les titres enfants ne sont pas fermés avant d'ouvrir
+												let grandChildrenList = getSubtotalTitleChilds($childLine, true); // renvoi la liste des id des enfants
+
+												if($childLine.attr('data-folder-status') == "closed"){
+													doNotDisplayLines = doNotDisplayLines.concat(grandChildrenList);
+												}
+												else if(o.config.closeMode == 'keepTitle' && $childLine.attr('data-folder-status') == "open"){
+													doNotHiddeLines = doNotDisplayLines.concat(grandChildrenList);
+												}
+											}
+
+											if (toggleStatus == 'closed') {
+												if(o.config.closeMode == 'keepTitle' && ($childLine.attr('data-issubtotal') == "title" || $childLine.attr('data-issubtotal') == "subtotal"  )){
+													$childLine.show();
+												}else if(!doNotHiddeLines.includes(childLineId)){
+													$childLine.hide();
+												}
+											} else {
+												if(!doNotDisplayLines.includes(childLineId)){
+													$childLine.show();
+												}
+											}
+
+											if(!$childLine.is(":visible")){
+												totalHiddenLines++;
+											}
+										});
 									}
 
-									//on récupère les lignes concernées en fonction du titre
-									$.ajax({
-										url: '<?php echo dol_buildpath('/subtotal/script/interface.php', 1); ?>'
-										,type: 'POST'
-										,data: {
-											json:1
-											,get: 'getLinesFromTitle'
-											,element: '<?php echo $element; ?>'
-											,elementid: '<?php echo $id; ?>'
-											,lineid: id_line_title
-										}
-									}).done(function(data) {
+									$collapseInfos.html('('+ totalHiddenLines +')');
+									if(totalHiddenLines == 0){
+										$collapseInfos.html('');
+									}
 
-										var data = JSON.parse(data);
+									// Mise à jour des parents pour l'affichage du nombre de lignes cachées
+									o.updateHiddenLinesCountInfoForParentTitles($titleLine);
 
-										//en fonction des lignes concernées, on définit les #id (html) concernée
-										$.each(data, function (title, tlines) {
+									if(toggleStatus == 'closed') {
+										$collapseBtn.html(o.config.img_folder_closed);
+										$collapseSimpleBtn.attr('title', o.config.langs.Subtotal_Show);
+										$collapseAllBtn.attr('title', o.config.langs.Subtotal_ForceShowAll);
+									}else{
+										$collapseBtn.html(o.config.img_folder_open);
+										$collapseSimpleBtn.attr('title', o.config.langs.Subtotal_Hide);
+										$collapseAllBtn.attr('title', o.config.langs.Subtotal_ForceHideAll);
+									}
 
-											if(action_hide == 0 && title !== id_line_title){
-												//cas où le bloc fermé et contient des sous-blocs : si on ouvre le bloc parent, on ouvre pas les sous-blocs
-											} else {
-												$.each(tlines, function (key, line) {
-													//si le dossier est fermé, et qu'on clique, alors on ouvre le bloc
-													if(action_hide == 0) {
-														//pour chaque #id (html) concerné, on affiche la ligne
-														$('#row-' + line).show();
-
-														//on met à jour l'extrafield "hideblock" pour indiquer que le bloc de cette ligne doit être affiché
-														$.ajax({
-															url: '<?php echo dol_buildpath('/subtotal/script/interface.php', 1); ?>'
-															, type: 'POST'
-															, data: {
-																json: 1
-																, set: 'update_hideblock_data'
-																, lineid: title
-																, element: '<?php echo $element; ?>'
-																, elementid: '<?php echo $id; ?>'
-																, value: 0
-															}
-														})
-
-														//On met à jour l'icône "dossier"
-														$('#collapse-' + title).html('<?php echo dol_escape_js(img_picto('', 'folder-open')); ?>');
-													}
-													//si le dossier est ouvert, et qu'on clique, alors on ferme le bloc
-													else {
-														//pour chaque #id (html) concerné, on affiche la ligne
-														$('#row-' + line).hide();
-
-														//on met à jour l'extrafield "hideblock" pour indiquer que le bloc de cette ligne doit être caché
-														$.ajax({
-															url: '<?php echo dol_buildpath('/subtotal/script/interface.php', 1); ?>'
-															, type: 'POST'
-															, data: {
-																json: 1
-																, set: 'update_hideblock_data'
-																, lineid: title
-																, element: '<?php echo $element; ?>'
-																, elementid: '<?php echo $id; ?>'
-																, value: 1
-															}
-														})
-
-														//On met à jour l'icône "dossier"
-														$('#collapse-' + title).html('<?php echo dol_escape_js(img_picto('', 'folder')); ?>');
-													}
-												});
-											}
-										});
-
-									});
+									// Si pas de titre pas besoin d'afficher le bouton dossier rouge
+									if(haveTitle){
+										$collapseAllBtn.show();
+									}else{
+										$collapseAllBtn.hide();
+									}
 								}
+							}
 
-
-								//Fonction qui permet d'ajouter l'option "Cacher les lignes" ou "Afficher les lignes"
-								function allFolderManage(){
-
-									//On ajoute une ligne au début du tableau qui propose de cacher toutes les lignes ou des les afficher
-									$('#tablelines>tbody:first').prepend('<tr><td colspan="100%" style="  text-align:right "><a id="hide_all" href="#"><?php echo img_picto('', 'folder-open', 'class="paddingright"').$langs->trans("Subtotal_HideAll"); ?></a>&nbsp;<a id="show_all" href="#"><?php echo img_picto('', 'folder', 'class="paddingright"').$langs->trans("Subtotal_ShowAll") ?></a></td></tr>')
-
-									//Lorsqu'on clique sur "cacher les lignes"
-									$("#hide_all").click(function() {
-										//on cache toutes les lignes de l'objet sauf les titres et sous-totaux
-										$("tr[data-product_type='0']").hide();
-										$("tr[data-product_type='1']").hide();
-
-										//on change tous les icônes "dossier" en fermé
-										$("[id*=collapse-]").html('<?php echo dol_escape_js(img_picto('', 'folder')); ?>');
-
-										//on sauvegarde l'information "hideblock" pour toutes les sections de l'objet
-										$.ajax({
-											url: '<?php echo dol_buildpath('/subtotal/script/interface.php', 1); ?>'
-											, type: 'POST'
-											, data: {
-												json: 1
-												, set: 'updateall_hideblock_data'
-												, element: '<?php echo $element; ?>'
-												, elementid: '<?php echo $id; ?>'
-												, value: 1
-											}
-										})
-										return false;
-									});
-
-									//Lorsqu'on clique sur "afficher les lignes"
-									$("#show_all").click(function() {
-										//on affiche toutes les lignes de l'objet sauf les titres et sous-totaux
-										$("tr[data-product_type='0']").show();
-										$("tr[data-product_type='1']").show();
-
-										//on change tous les icônes "dossier" en ouvert
-										$("[id*=collapse-]").html('<?php echo dol_escape_js(img_picto('', 'folder-open')); ?>');
-
-										//on sauvegarde l'information "hideblock" pour toutes les sections de l'objet
-										$.ajax({
-											url: '<?php echo dol_buildpath('/subtotal/script/interface.php', 1); ?>'
-											, type: 'POST'
-											, data: {
-												json: 1
-												, set: 'updateall_hideblock_data'
-												, element: '<?php echo $element; ?>'
-												, elementid: '<?php echo $id; ?>'
-												, value: 0
-											}
-										})
-
-										return false;
-									});
+							// initialisation des lignes affichées ou non
+							$('tr[data-issubtotal="title"]').each(function() {
+								let lineId = $( this ).attr('data-id');
+								if(lineId != null){
+									if(o.config.linesToHide.includes(lineId)){
+										o.toggleChildFolderStatusDisplay(lineId, 'closed');
+									}else{
+										o.toggleChildFolderStatusDisplay(lineId, 'open');
+									}
 								}
-
 							});
-						</script>
+
+							// Lors du clic sur un dossier, on cache ou faire apparaitre les lignes contenues dans le bloc concerné
+							$(document).on("click",".fold-subtotal-btn",function(event) {
+								event.preventDefault();
+								let targetTitleLineId = $(this).attr('data-title-line-target');
+								if(targetTitleLineId != undefined){
+									// folderManage_click(targetTitleLineId);
+									let titleRow = $('#row-' + targetTitleLineId);
+									let newStatus = titleRow.attr('data-folder-status') == 'closed' ? 'open' : 'closed'
+									let sendData = {
+										element : o.config.element,
+										element_id : o.config.element_id,
+										titleStatusList : [{
+											'id': targetTitleLineId,
+											'status': newStatus !== 'closed' ? 0 : 1,
+										}]
+									};
+
+									/**
+									 * Pour les boutons de type "block" bouton pour ouvrir / fermer tous les blocs enfants (ex dossier rouge)
+									 **/
+									if($(this).attr('data-toggle-all-children') == '1'){ //o.config.closeMode == 'keepTitle'
+										let childrenList = getSubtotalTitleChilds(titleRow, true); // renvoi la liste des id des enfants
+										if(childrenList.length > 0) {
+											childrenList.forEach((childLineId) => {
+												let $childLine = $('#'+childLineId);
+												if ($childLine.attr('data-issubtotal') == "title"){
+													sendData.titleStatusList.push({
+														'id': $childLine.attr('data-id'),
+														'status': newStatus !== 'closed' ? 0 : 1,
+													});
+													o.toggleChildFolderStatusDisplay($childLine.attr('data-id'), newStatus);
+												}
+											});
+										}
+									}
+
+									o.toggleChildFolderStatusDisplay(targetTitleLineId, newStatus); // devrait être dans le callback ajax success mais pour plus d'ergonomie et rapidité de feedback je le sort
+									o.callInterface('set' , 'update_hideblock_data', sendData, function(response){
+										// TODO gérer un retour en cas d'érreur
+										// o.toggleChildFolderStatusDisplay(targetTitleLineId, newStatus);
+									})
+								}
+							});
+
+
+							//Fonction qui permet d'ajouter l'option "Cacher les lignes" ou "Afficher les lignes"
+							$('#tablelines>tbody:first').prepend(
+								'<tr>' +
+								'	<td colspan="100%" style="  text-align:right ">' +
+								'		<span id="hide_all"  class="toggle-all-folder-status" data-folder-status="closed" >'+o.config.img_folder_open+'&nbsp;'+o.config.langs.Subtotal_HideAll+'</span>' +
+								'		&nbsp;' +
+								'		<span id="show_all" class="toggle-all-folder-status" data-folder-status="open"  >'+o.config.img_folder_closed + '&nbsp;'+o.config.langs.Subtotal_ShowAll+'</span>' +
+								'	</td>' +
+								'</tr>'
+							);
+
+
+							// Lors du clic sur un dossier, on cache ou faire apparaitre les lignes contenues dans le bloc concerné
+							$(document).on("click",".toggle-all-folder-status",function(event) {
+								event.preventDefault();
+								newStatus = $( this ).attr('data-folder-status');
+								$( this ).fadeOut();
+
+								let sendData = {
+									element : o.config.element,
+									element_id : o.config.element_id,
+									titleStatusList : []
+								};
+
+								$('#tablelines tr[data-issubtotal=title]').each(function( index ) {
+									sendData.titleStatusList.push({
+										'id': $( this ).attr('data-id'),
+										'status': newStatus !== 'closed' ? 0 : 1,
+									});
+
+									//TODO manage response feedback to rollback display on error
+									o.toggleChildFolderStatusDisplay($( this ).attr('data-id'), newStatus);
+								});
+
+								o.callInterface('set' , 'update_hideblock_data', sendData, function(response){
+									// $('#tablelines tr[data-issubtotal=title]').each(function( index ) {
+									// 	//TODO manage response feedback
+									// });
+								});
+
+
+								$( this ).fadeIn();
+							});
+
+
+							o.checkListOfLinesIdHaveTitle = function(childrenList){
+								if(!Array.isArray(childrenList)){
+									return false;
+								}
+
+								childrenList.forEach((childLineId) => {
+									let $childLine = $('#' + childLineId);
+									if ($childLine.length > 0 && $childLine.attr('data-issubtotal') == "title") {
+										return true;
+									}
+								});
+
+								return false;
+							}
+
+							/**
+							 *
+							 * @param {string} typeAction
+							 * @param {string} action
+							 * @param sendData
+							 * @param callBackFunction
+							 */
+							o.callInterface = function ( typeAction = 'get' , action, sendData = {}, callBackFunction){
+
+								let ajaxData = {
+									'data': sendData,
+									'token': o.newToken,
+								};
+
+								if(typeAction == 'set'){
+									ajaxData.set = action;
+								}else{
+									ajaxData.get = action;
+								}
+
+								$.ajax({
+									method: 'POST',
+									url: o.config.interfaceUrl,
+									dataType: 'json',
+									data: ajaxData,
+									success: function (response) {
+
+										if (typeof callBackFunction === 'function'){
+											callBackFunction(response);
+										} else {
+											console.error('Callback function invalide for callKanbanInterface');
+										}
+
+										if(response.newToken != undefined){
+											o.newToken = response.newToken;
+										}
+
+										if(response.msg.length > 0) {
+											o.setEventMessage(response.msg, response.result > 0 ? true : false, response.result == 0 ? true : false );
+										}
+									},
+									error: function (err) {
+
+										if(err.responseText.length > 0){
+
+											// detect login page in case of just disconnected
+											let loginPage = $(err.responseText).find('[name="actionlogin"]');
+											if(loginPage != undefined && loginPage.val() == 'login'){
+												o.setEventMessage(o.langs.errorAjaxCallDisconnected, false);
+
+												setTimeout(function (){
+													location.reload();
+												}, 2000);
+
+											}else{
+												o.setEventMessage(o.langs.errorAjaxCall, false);
+											}
+										}
+										else{
+											o.setEventMessage(o.langs.errorAjaxCall, false);
+										}
+									}
+								});
+							}
+
+
+							/**
+							 *
+							 * @param {string} msg
+							 * @param {boolean} status
+							 * @param {boolean} sticky
+							 */
+							o.setEventMessage = function (msg, status = true, sticky = false){
+
+								let jnotifyConf = {
+									delay: 1500                               // the default time to show each notification (in milliseconds)
+									, type : 'error'
+									, sticky: sticky                             // determines if the message should be considered "sticky" (user must manually close notification)
+									, closeLabel: "&times;"                     // the HTML to use for the "Close" link
+									, showClose: true                           // determines if the "Close" link should be shown if notification is also sticky
+									, fadeSpeed: 150                           // the speed to fade messages out (in milliseconds)
+									, slideSpeed: 250                           // the speed used to slide messages out (in milliseconds)
+								}
+
+
+								if(msg.length > 0){
+									if(status){
+										jnotifyConf.type = '';
+										$.jnotify(msg, jnotifyConf);
+									}
+									else{
+										$.jnotify(msg, jnotifyConf);
+									}
+								}
+								else{
+									$.jnotify('ErrorMessageEmpty', jnotifyConf);
+								}
+							}
+
+						})(subtotalFolders);
+
+					});
+				</script>
+
 				<?php
-
-
 			}
-			/**Gestion des dossiers qui permettent de réduire un bloc**/
 
 	}
 }