+++ ChangeLog.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 # [Unreleased]
 
+- NEW : Add new option to chose for folder management behavior  - *17/07/2023* - 3.22.0
+  Also add a new BTN to open/close a complete block with its children too.
+- FIX : Folder management ajax call - *10/07/2023* - 3.21.1
 - NEW : Ajout des marges par blocs - *28/06/2023* - 3.21.0
 - NEW : add folder management - *28/06/2023* - 3.20.0